refactor(settings): move forms to widget
