+++ src/entities/settings/index.ts
@@ -1,2 +1,14 @@
-export * from './config';
-export * from './model';
+export { TimerPresetsConfig } from './config';
+export type {
+  INotificationSettings,
+  ISoundSettings,
+  IThemeSettings,
+  ITimerSettings,
+} from './model';
+export {
+  resetSettings,
+  useNotifySettingsStore,
+  useSoundSettingsStore,
+  useThemeSettingsStore,
+  useTimerSettingsStore,
+} from './model';
