+++ src/entities/sound/index.ts
@@ -1,1 +1,1 @@
-export * from './ui';
+export { SoundSelect } from './ui';
