+++ src/pages/main/ui/main-page.tsx
@@ -1,0 +1,7 @@
+import { SettingsDialog } from '@/widgets/settings-dialog';
+
+const MainPage = () => {
+  return <SettingsDialog />;
+};
+
+export { MainPage };
