+++ src/widgets/settings-dialog/ui/forms/notify-settings-form/config/index.ts
