+++ src/widgets/settings-dialog/ui/forms/notify-settings-form/config/notify-status-descriptions.ts
