+++ src/widgets/settings-dialog/ui/forms/themes-settings-form/config/index.ts
