+++ src/widgets/settings-dialog/ui/forms/themes-settings-form/config/preview-list.ts
