+++ src/widgets/settings-dialog/ui/forms/themes-settings-form/model/type.ts
