+++ src/widgets/settings-dialog/ui/forms/timer-settings-form/config/form-limits.ts
