+++ src/widgets/settings-dialog/ui/forms/timer-settings-form/config/index.ts
