+++ src/widgets/settings-dialog/ui/forms/timer-settings-form/index.ts
