+++ src/widgets/settings-dialog/ui/forms/timer-settings-form/lib/format-number.ts
