+++ src/widgets/settings-dialog/ui/forms/timer-settings-form/lib/mask-number.ts
