+++ src/widgets/settings-dialog/ui/forms/timer-settings-form/ui/number-input.tsx
